platform/x86/amd/pmc: Add 10ms delay to 2nd s0i3 cycle for all APUs
During s0i3 if a user unplugs the power adapter the EC may throttle
the APU (like a thermal event would) and wakes the system from hardware
sleep.

As only one interrupt is active (for the ACPI SCI) the system will go
back into hardware sleep.  This sequence races with the hardware and
the net outcome is that the system will remain throttled when a second
interrupt is active actually waking the system from suspend.

Reported-by: aros@gmx.com
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=218305
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>